experts/data_engineer/pyproject.toml
[project]
name = "data_engineer"
version = "0.1.0"
description = "Data Engineer Agent for Smart Task Hub"
dependencies = [
    "google-adk[a2a]>=1.25.0",
]

[tool.setuptools]
py-modules = ["agent", "__init__"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
